File size limit is 5TB with Drive client
According to FAQ: https://aws.amazon.com/workdocs/faq/ (General section):

Q. What is the maximum file size that can be uploaded to WorkDocs?
Using the Drive client on Mac or PC, you can upload individual files up to 5TB in size. We recommend using the Drive client to transfer large files to WorkDocs.

I've tested and be able to successfully transfer a 10GB file with the Mac Drive client.